+++ Tests/PeripheryTests/Extensions/FilePathTest.swift
@@ -3,6 +3,13 @@ import XCTest
 import SystemPackage
 
 class FilePathTest: XCTestCase {
+    func testMakeAbsolute() {
+        let current = FilePath("/current")
+        XCTAssertEqual(FilePath.makeAbsolute("/a", relativeTo: current).string, "/a")
+        XCTAssertEqual(FilePath.makeAbsolute("a", relativeTo: current).string, "/current/a")
+        XCTAssertEqual(FilePath.makeAbsolute("./a", relativeTo: current).string, "/current/a")
+    }
+
     func testRelativeTo() {
         XCTAssertEqual(FilePath("/a/b/c").relativeTo(FilePath("/a/b/c")).string, ".")
         XCTAssertEqual(FilePath("/a/b/c/d").relativeTo(FilePath("/a/b")).string, "c/d")